+++ masterJenisAset/index.php
@@ -30,7 +30,7 @@ $result = mysqli_query($connection, "SELECT * FROM masterjenisaset");
                   <tr>
                     <td><?= $data['namaJenisAset'] ?></td>
                     <td>
-                      <a class="btn btn-sm btn-danger mb-md-0 mb-1" href="delete.php?id_jenisAset=<?= $data['id_jenisAset'] ?>">
+                      <a class="btn btn-sm btn-danger mb-md-0 mb-1" href="delete.php?id_jenisAset=<?= $data['id_jenisAset'] ?>" onclick="return confirm('Anda yakin ingin menghapus data ini?');">
                         <i class="fas fa-trash fa-fw"></i>
                       </a>
                       <a class="btn btn-sm btn-info" href="edit.php?id_jenisAset=<?= $data['id_jenisAset'] ?>">
