+++ masterLokasi/index.php
@@ -31,7 +31,7 @@ $result = mysqli_query($connection, "SELECT * FROM masterlokasi");
                   <tr>
                     <td><?= $data['namaLokasi'] ?></td>
                     <td>
-                      <a class="btn btn-sm btn-danger mb-md-0 mb-1" href="delete.php?id_lokasi=<?= $data['id_lokasi'] ?>">
+                      <a class="btn btn-sm btn-danger mb-md-0 mb-1" href="delete.php?id_lokasi=<?= $data['id_lokasi'] ?>" onclick="return confirm('Anda yakin ingin menghapus data ini?');">
                         <i class="fas fa-trash fa-fw"></i>
                       </a>
                       <a class="btn btn-sm btn-info" href="edit.php?id_lokasi=<?= $data['id_lokasi'] ?>">
