+++ masterMaintenance/index.php
@@ -40,7 +40,7 @@ $result = mysqli_query($connection, "SELECT * FROM datamt");
                     <td><?= $data['keterangan'] ?></td>
                     <td><?= $data['status'] ?></td>
                     <td>
-                      <a class="btn btn-sm btn-danger mb-md-0 mb-1" href="delete.php?id_mt=<?= $data['id_mt'] ?>">
+                      <a class="btn btn-sm btn-danger mb-md-0 mb-1" href="delete.php?id_mt=<?= $data['id_mt'] ?>" onclick="return confirm('Anda yakin ingin menghapus data ini?');">
                         <i class="fas fa-trash fa-fw"></i>
                       </a>
                       <a class="btn btn-sm btn-info" href="edit.php?id_mt=<?= $data['id_mt'] ?>">
